// packages/battle/src/v2/schemas/context.schema.d.ts
export declare const UseSkillContextSchema: import("@sinclair/typebox").TObject<{
    type: import("@sinclair/typebox").TLiteral<"use-skill">;
    parentId: import("@sinclair/typebox").TString;
    petId: import("@sinclair/typebox").TString;
    skillId: import("@sinclair/typebox").TString;
    originPlayerId: import("@sinclair/typebox").TString;
    selectTarget: import("@sinclair/typebox").TUnion<import("@sinclair/typebox").TLiteral<string>[]>;
    priority: import("@sinclair/typebox").TNumber;
    category: import("@sinclair/typebox").TUnion<import("@sinclair/typebox").TLiteral<string>[]>;
    element: import("@sinclair/typebox").TString;
    power: import("@sinclair/typebox").TNumber;
    accuracy: import("@sinclair/typebox").TNumber;
    petAccuracy: import("@sinclair/typebox").TNumber;
    rage: import("@sinclair/typebox").TNumber;
    evasion: import("@sinclair/typebox").TNumber;
    critRate: import("@sinclair/typebox").TNumber;
    ignoreShield: import("@sinclair/typebox").TBoolean;
    ignoreStageStrategy: import("@sinclair/typebox").TUnion<import("@sinclair/typebox").TLiteral<string>[]>;
    multihit: import("@sinclair/typebox").TUnion<[import("@sinclair/typebox").TTuple<[import("@sinclair/typebox").TNumber, import("@sinclair/typebox").TNumber]>, import("@sinclair/typebox").TNumber]>;
    available: import("@sinclair/typebox").TBoolean;
    actualTargetId: import("@sinclair/typebox").TOptional<import("@sinclair/typebox").TString>;
    hitResult: import("@sinclair/typebox").TBoolean;
    crit: import("@sinclair/typebox").TBoolean;
    multihitResult: import("@sinclair/typebox").TNumber;
    currentHitCount: import("@sinclair/typebox").TNumber;
    damageType: import("@sinclair/typebox").TUnion<import("@sinclair/typebox").TLiteral<string>[]>;
    typeMultiplier: import("@sinclair/typebox").TNumber;
    stabMultiplier: import("@sinclair/typebox").TNumber;
    critMultiplier: import("@sinclair/typebox").TNumber;
    baseDamage: import("@sinclair/typebox").TNumber;
    randomFactor: import("@sinclair/typebox").TNumber;
    defeated: import("@sinclair/typebox").TBoolean;
}>;
export declare const DamageContextSchema: import("@sinclair/typebox").TObject<{
    type: import("@sinclair/typebox").TLiteral<"damage">;
    parentId: import("@sinclair/typebox").TString;
    sourceId: import("@sinclair/typebox").TString;
    targetId: import("@sinclair/typebox").TString;
    baseDamage: import("@sinclair/typebox").TNumber;
    damageType: import("@sinclair/typebox").TUnion<import("@sinclair/typebox").TLiteral<string>[]>;
    crit: import("@sinclair/typebox").TBoolean;
    effectiveness: import("@sinclair/typebox").TNumber;
    ignoreShield: import("@sinclair/typebox").TBoolean;
    randomFactor: import("@sinclair/typebox").TNumber;
    modified: import("@sinclair/typebox").TTuple<[import("@sinclair/typebox").TNumber, import("@sinclair/typebox").TNumber]>;
    minThreshold: import("@sinclair/typebox").TNumber;
    maxThreshold: import("@sinclair/typebox").TNumber;
    damageResult: import("@sinclair/typebox").TNumber;
    available: import("@sinclair/typebox").TBoolean;
    element: import("@sinclair/typebox").TOptional<import("@sinclair/typebox").TString>;
}>;
// ...
export type UseSkillContextData = Static<typeof UseSkillContextSchema>;
export type DamageContextData = Static<typeof DamageContextSchema>;